+++ projects/Gibbed.Volition.Packing/VPP/Unpacker.cs
@@ -228,7 +228,7 @@ namespace Gibbed.Volition.Packing.VPP
                                 }
                                 else
                                 {
-                                    using (var temp = data.ReadToMemoryStream(entry.CompressedSize))
+                                    using (var temp = data.ReadToMemoryStream((int)entry.CompressedSize))
                                     {
                                         var zlib = new InflaterInputStream(temp);
                                         output.WriteFromStream(zlib, entry.UncompressedSize);
